renamed the src/external folder to src/scripts
